fixed warning

+++ src/util/indent.cpp
@@ -67,13 +67,13 @@ void indentLine(State* state) {
 void indentRange(State* state) {
     int indentDifference = getIndentLevel(state, state->row) - getNumLeadingSpaces(state->data[state->row]);
     if (indentDifference > 0) {
-        for (int i = state->row; i <= state->visual.row; i++) {
+        for (int i = state->row; i <= (int) state->visual.row; i++) {
             for (int j = 0; j < indentDifference; j++) {
                 state->data[i] = ' ' + state->data[i];
             }
         }
     } else if (indentDifference < 0) {
-        for (int i = state->row; i <= state->visual.row; i++) {
+        for (int i = state->row; i <= (int) state->visual.row; i++) {
             state->data[i] = state->data[i].substr(-1 * indentDifference);
         }
     }
